se agrego la parte de eliminar el mensaje
agregue el endpoint de eliminar el mensaje y mejore el tema del controlador y el modelo del mensaje para que sea mas facil de menejar.

+++ backend/src/Controllers/MessageController.php
@@ -2,7 +2,8 @@
     namesPace App\Controllers;
 
     use App\Models\MessageModel;
-    use Psr\Http\Message\ResponseInterface as Response;
+use App\Models\UserModel;
+use Psr\Http\Message\ResponseInterface as Response;
     use Psr\Http\Message\ServerRequestInterface as Request;
 
     class MessageController{
@@ -13,13 +14,34 @@
             if($message_id == 0 ){
                 return $this->mensaje($response, "El mensaje a eliminar no existe", 400);
             }
+            //obtenemos el usuario
+            $usuario = $request->getAttribute('usuario');
+            //Verificamos si es admin
+            $isAdmin = UserModel::EsAdmin($usuario);
 
-            $result = MessageModel::deleteMessage($message_id);
-            if($result){
-                return $this->mensaje($response, "Mensaje eliminado con exito.", 200);
-            }else{
-                return $this->mensaje($response, "Error al querer eliminarl el mensaje.", 500);
+            if($isAdmin){
+                $result = MessageModel::deleteMessage($message_id);
+                if($result){
+                    return $this->mensaje($response, "Mensaje eliminado con exito.", 200);
+                }else{
+                    return $this->mensaje($response, "El mensaje no existe o ya fue eliminado.", 409);
+                }
             }
+
+            //obtenemos el mensaje para luego comparar si el usuario es el creador del mensaje 
+            $message = MessageModel::getMessage($message_id);
+            if($message['enviado_por'] == $usuario){
+                $result = MessageModel::deleteMessage($message_id);
+                if($result){
+                    return $this->mensaje($response, "Mensaje eliminado con exito.", 200);
+                }else{
+                    return $this->mensaje($response, "El mensaje no existe o ya fue eliminado.", 409);
+                }
+            }
+            else{
+                return $this->mensaje($response, "Error al querer eliminar el mensaje. El mesanje a eliminar no es suyo", 409);
+            }
+            
     }
 
     private function mensaje($response, $msj, $num): Response {
